requirements.txt
discord.py==2.0.0
aiohttp==3.8.1
asyncio==3.4.3
logging==0.5.1.2
